show how to use QBasicTimer, never forget "self"
